[WIP] Core: Global module hooks
Before this change, the global module object was not created by
`createModule()`. Instead, it was declared in config.js as the first
value for `config.currentModule`, with only some of the properties,
enough to satisfy internal execution but none of the public stuff
like moduleId, testEnvironment, etc.

Since 2017 with commit a53acdf2a this did include a `hooks` object.
However, this was only defined to satisfy `runHook()` in test.js, so
that is executes a global test (which points to this incomplete module
object as its parent), it can pretend it's a normal module with a valid
but always-empty hooks object that we didn't expose a way to add
functions to. And that's good, because if we did expose it before,
it wouldn't have worked reliably.

== Changes ==

Only two small changes needed to support global hooks

1) Actually execute the global module.

The global module was introduced in 2015 with commit 168b048c682, and
we decided to only execute it when at least one global test is defined.
I've retroactively documented the reason for this, and propose to
change this in QUnit 3.

2) Balance it correctly.

Before this change, if there was one or more global tests defined,
we did schedule and execute the global module, but it wasn't
associated anywhere as `parentModule` and also wasn't on the
`moduleStack`. As such, it was effectively a standalone module sibling
to other top-level modules, which was an unobservable detail until
now, but if kept would have caused the beforeEach and afterEach hooks
to not be inherited, and the `after` hook to execute as soon as
the next module begins, instead of at the very end of the test run.

== Bonus ==

As a bonus, this change resolves one of the rollup warnings about
circular dependencies. Previously, core.js > module.js > core.js, due
to `globalSuite` being defined in core.js. I've moved this down to
module.js.