CLI: Apply trace cleaning and internal greying to assert failure traces
Follows-up https://github.com/qunitjs/qunit/pull/1789, which applied
this to traces under uncaught errors. We now apply the same to traces
under assertion failures as well.

(TODO: Split into separate commit)

Omit noisy and confusing actual:null,expected:undefined from TAP
output. This is prevented in HtmlReporter due to hasOwn check, but
this check fails in TAP because it uses testEnd.errors instead of
QUnit.log(). And, in Test.js#logAssertion, where originally expected
is allowed to be unset, it gets forged in the call to testReport
in order to satisfy the API. Indeed, js-reporters never allowed for
unset actual/expected.

Fix this by instead deciding to skip rendering if both are strictly
equal to the value of undefined, which I believe cannot be useful
information. However, for that to be true, we have to change
`actual:null` to `actual:undefined` in pushFailure(). Otherwise,
it can easily be a genuine diff.